i175 Set the parent project to hibernate-search-parent and take advantage of this
This removes a bunch of dependency management entries, properties, etc.
because they are already provided in hibernate-search-parent.

Other changes:

 * Some forbidden uses of locale-sensitive methods had to be adapted
 * WildFly deployment structure files had to be adapted to support
   Hibernate ORM 5.2.7
 * The build-config project was removed (we use Hibernate Search's)

+++ core/src/main/java/org/hibernate/search/jsr352/internal/steps/lucene/EntityReader.java
@@ -177,7 +177,7 @@ public class EntityReader extends AbstractItemReader {
 
 		final int partitionId = Integer.parseInt( partitionIdStr );
 
-		LOGGER.debugf( "[partitionId=%d] open reader for entity %s ...", partitionId, entityName );
+		LOGGER.debugf( "[partitionId=%d] open reader for entity %s ...", (Integer) partitionId, entityName );
 		JobContextData jobData = (JobContextData) jobContext.getTransientUserData();
 		entityType = jobData.getIndexedType( entityName );
 		PartitionBound bound = jobData.getPartitionBound( partitionId );
